move HTTP file stream copy logic to Squidlr.Api
